requirements.txt
yfinance==1.2.0
prophet
pandas
python-dotenv
requests
scikit-learn
pandas-ta
matplotlib
seaborn
nltk
python-telegram-bot
